+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+- Added extension `BenchmarkHelper`.
+- Improved calculation of sample size.
+- Extended docs, added image.
+- Added class `Score`.
+
 ## 0.1.0
 - Added tests for class `Stats`.
 - `ColorEmitter` now scales benchmark score stats (that are larger than 1000 us).